refactor: unify header across journal and management
